[REFACTOR] 패키지구조 변경 - Firebase 패키지 external로 이동, Announcemnet, Attendance, Memo, Register, Session 각 도메인별 Controller, Service, Repository 패키지 변경
