+++ Swith-Server/src/main/java/com/swith/domain/interest/repository/InterestRepository.java
@@ -1,0 +1,7 @@
+package com.swith.domain.interest.repository;
+
+import com.swith.domain.interest.entity.Interest;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InterestRepository extends JpaRepository<Interest, Integer> {
+}
