+++ Swith-Server/src/main/java/com/swith/external/firebase/GroupInfoFcm/model/GroupInfoFcmMessage.java
@@ -1,0 +1,4 @@
+package com.swith.external.firebase.GroupInfoFcm.model;
+
+public class GroupInfoFcmMessage {
+}
