+++ Swith-Server/src/main/java/com/swith/external/firebase/GroupInfoFcm/model/GroupInfoFcmResponse.java
@@ -1,0 +1,4 @@
+package com.swith.external.firebase.GroupInfoFcm.model;
+
+public class GroupInfoFcmResponse {
+}
