+++ Swith-Server/src/main/java/com/swith/external/firebase/model/FcmRequest.java
@@ -1,0 +1,15 @@
+package com.swith.external.firebase.model;
+
+
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class FcmRequest {
+    private String token;
+    private String title;
+    private String body;
+}
