+++ Swith-Server/src/main/java/com/swith/external/firebase/model/FcmResponse.java
@@ -1,0 +1,16 @@
+package com.swith.external.firebase.model;
+
+
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class FcmResponse {
+    private String title;
+    private String body;
+    private String android_channel_id;
+    private String channel_id;
+
+}
