--- Swith-Server/src/main/java/com/swith/src/firebase/GroupInfoFcm/model/GroupInfoFcmMessage.java
@@ -1,4 +1,0 @@
-package com.swith.src.firebase.GroupInfoFcm.model;
-
-public class GroupInfoFcmMessage {
-}
