--- Swith-Server/src/main/java/com/swith/src/firebase/GroupInfoFcm/model/GroupInfoFcmResponse.java
@@ -1,4 +1,0 @@
-package com.swith.src.firebase.GroupInfoFcm.model;
-
-public class GroupInfoFcmResponse {
-}
