--- Swith-Server/src/main/java/com/swith/src/firebase/model/FcmRequest.java
@@ -1,15 +1,0 @@
-package com.swith.src.firebase.model;
-
-
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class FcmRequest {
-    private String token;
-    private String title;
-    private String body;
-}
