--- Swith-Server/src/main/java/com/swith/src/firebase/model/FcmResponse.java
@@ -1,16 +1,0 @@
-package com.swith.src.firebase.model;
-
-
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class FcmResponse {
-    private String title;
-    private String body;
-    private String android_channel_id;
-    private String channel_id;
-
-}
