--- Swith-Server/src/main/java/com/swith/src/repository/InterestRepository.java
@@ -1,7 +1,0 @@
-package com.swith.src.repository;
-
-import com.swith.domain.interest.entity.Interest;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface InterestRepository extends JpaRepository<Interest, Integer> {
-}
